feat: forbid f32 and f64 as map key type
According to IEEE-754 2019:
NaN (Not a Number) is not equal to any value, including itself.

This inequality leads to different implementations:

1.  Implementations that "follows" the standard

    In this case, an unbounded number of values may be bind to the key NaN
    and cannot be accessed.

    This is the implementation chosen by Golang.

2.  Implementations that normalize NaNs and consider that NaN
    is equal to itself

    This is the implementation chosen by JavaScript

3.  Implementations that rely on the binary comparison of NaNs

This makes complex the support of f32 and f64 as map key type.

To avoid this complexity the ongoing BARE draft forbids their usage as
map key type.

BREAKING CHANGES: floats were previosuly accepted as map key type.