test(lint): use denolint
Note that we also lint tests-corpus to ensure that
bare-ts generates high-quality codes.

For now, we do not use the config file because
it does not take into account the field `files`.
I opened an issue for that:

  https://github.com/napi-rs/node-rs/issues/635

I added a config files for deno lint command.
This allows to manually check extra rules:

  deno lint --config .deno.json

Note that some rules are violated.
I excluded scripts and tests-corpus to avoid excessive rule vilation.
These two problems come from bugs that are reported:

  https://github.com/denoland/deno_lint/issues/1018
  https://github.com/denoland/deno_lint/issues/1048
  https://github.com/denoland/deno_lint/issues/1049