chore: emit ES2020 for smaller builds
bare-ts now publishes ES2020 builds.
This allows to output smaller builds.
This should cause no issue since we require a node version
^14.18 or >=16.

Note that a single construct is not supported:
spread-parameters-after-optional-chaining
See https://node.green/#ES2020

A more robust solution could be to rely on compat tables
for emiting builds.
I attempted to use the `--target=` command line option of ESbuild.
However, ESbuild seems to ignore it or so the compat table is flawed...