chore: avoid dual package hazard for bundlers
This change avoids dual package hazard for bundlers.
The issue is still present for node.

When an application is using a package that provides both CJS and ESM,
there is a risk of bugs if both versions of the package get loaded.
For instance, classes, singletons, ... are duplicated.

One way of avoiding the dual package hazard that works both for
bundlers and when running natively in node is to put all of your code
in the `require` condition as CJS and have the `import` condition
just be a light ESM wrapper that calls require on your package and
re-exports the package using ESM syntax.
Unfortunately, this approach doesn't provide good tree-shaking.
Indeed, a lot of bundlers, including ESbuild, do not tree-shake CJS.

Another way of avoiding a dual package hazard is to use the
bundler-specific `module` condition to direct bundlers to always load
the ESM version of your package while letting node always fallback to
the CJS version of your package.
Both `import` and `module` are intended to be used with ESM but unlike
`import`, `module` is always active even if the import path was loaded
using a require call.
This works well with bundlers because they support loading ESM using
require syntax, but it's not something that can work with node
because node deliberately doesn't implement loading ESM using require.

A tjird approach could be to publish ESM-only packages.
While we enforce use of a node version that supports ESM,
an application could transpile its code to CJS.
In this case, the application cannot load our module.

We adopt the second approach by adding `module` condition.
Eventually we plan t oadopt the third approach.
However, we think that the Node community is not yet ready for ESM-only
packages.