chore: use jsonc extension for commented tsconfig
TSC allows comments in `tsconfig.json` files.
Ideally the file should use the `jsonc` extension.
TypeScript team seems rejecting the idea of this support.
See [this issue](microsoft/TypeScript#43121).

By the way, `jsonc` extension can be used for tsconfig that are used
in `extends` of another _tsconfig_.

This fact was observed in this following thread:
https://www.typescriptlang.org/docs/handbook/tsconfig-json.html